+++ includes/ClicShopping/Apps/Customers/Reviews/Sites/ClicShoppingAdmin/Pages/Home/Actions/Reviews/Update.php
@@ -18,6 +18,7 @@ class Update extends \ClicShopping\OM\PagesActionsAbstract
   public function execute()
   {
     $CLICSHOPPING_Reviews = Registry::get('Reviews');
+    $CLICSHOPPING_Hooks = Registry::get('Hooks');
 
     $page = (isset($_GET['page']) && is_numeric($_GET['page'])) ? (int)$_GET['page'] : 1;
 
@@ -49,6 +50,8 @@ class Update extends \ClicShopping\OM\PagesActionsAbstract
       ];
 
       $CLICSHOPPING_Reviews->db->save('reviews_description', $sql_array, ['reviews_id' => (int)$reviews_id]);
+
+      $CLICSHOPPING_Hooks->call('Reviews', 'Update');
 
       $CLICSHOPPING_Reviews->redirect('Reviews&page=' . $page);
     }