+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 version 3.438
+Remove Davinci for Gpt3.5-turbo
+Url Rewiting improvement