+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 version 3.438
 Remove Davinci for Gpt3.5-turbo
 Url Rewiting improvement
-
+Recommendations refactored
 
 
 --------------------